git-svn-id: https://svn.code.sf.net/p/cspfj/code/trunk/cspfj@770 b3bf8218-cbbe-40ea-ae7b-cf512e877a40

+++ src/main/scala/cspfj/heuristic/DDegOnDom.scala
@@ -8,7 +8,7 @@ import cspfj.problem.Variable;
 final class DDegOnDom(val problem: Problem) extends VariableHeuristic with RandomBreak {
 
   def score(variable: Variable) =
-    dDeg(variable) / variable.dom.size
+    dDeg(variable).toDouble / variable.dom.size
 
   private def dDeg(variable: Variable) =
     variable.constraints.iterator.filter(!_.isEntailed).size
